Discover git repositories in parent directories
If the specified/default repository directory is not a valid git
repository, now we will traverse upwards in search of a git repository.

We then need to set `@options[:repo]` to the working directory in case
we did end up traversing upwards. It's used in `setup_rubocop` so if we
don't update it, that method will look for `.rubocop.yml` in the wrong
location.