Consolidate imports of `rubocop`
I think the original idea here was to delay requiring RuboCop so that if
we, for example, exit early due to invalid CLI arguments, we don't have
to wait unnecessarily for RuboCop to be imported. But I think that this
approach is more confusing and doesn't save that much time in the end.